Manually install hwasan runtime on SANITIZE_TARGET=hwaddress builds
Due to bugs in Soong around prebuilt libraries with stubs versions,
the hwasan runtime was installed into /system/lib64/bootstrap, but the
module was not given a ".bootstrap" suffix.  Once the bugs are fixed,
the hwasan runtime is no longer installed because the dependencies
on "libclang_rt.hwasan-aarch64-android" are now on an uninstallable
stubs library.  Add libclang_rt.hwasan-aarch64-android.bootstrap
to PRODUCT_PACKAGES when SANITIZE_TARGET=hwaddress is set.

Test: m checkbuild
Change-Id: I732d37505274c7bf804396921202f2bcca4484ec