Add bootclasspath config to its own makefile
This is part of adding some devices/products suitable for unbundled
builds. An unbundled product does not need all the PRODUCT_PACKAGES
etc present in the normal hierarchy of products, but does need to know
which jars are on the bootclasspath.

Therefore, factor out the bootclasspath config from base_system.mk and
put it in its own makefile where it can be inherited by these unbundled
products.

Bug: 172256440
Test: diff presubmit artifacts (noop)
Change-Id: Id3f91e387bba2e0525b4010f22fe380d1d8be537