Update the default compiler filter used at boot time
For the first-boot (happening after upgrades) use verify as quicken is deprecated.

For the usual boot operations use extract instead of the verify. Under normal behaviour all apps should already be optimized (either with the first-boot option, a/b option or in bg-dexopt). This will speed up the boot process in corner cases and reduce resource contention that could make the device janky in at boot time.

Bug: 172114559
Change-Id: I33f204911bdc825bf6731feb5cea7c260382798b
Test: m