Remove com.android.geotz jar from system server
This commit removes the com.android.geotz apex from the build.

This apex is associated with the "offline" location time zone provider
from the nascent com.android.geotz module, which has been configured as
the "secondary" provider. This module will not to be updatable in S so
needs to be removed before release.

This is effectively a revert of commit
725c59999cc4008dfc828844b5629ab1c484d845.

Bug: 178169250
Bug: 172546738
Test: treehugger
Change-Id: I734388383030984d5c3e7a1373a77da013a6cb8e