Remove use of REMOVE_ATB_FROM_BCP
This change cleans up after the work to remove the android.test.base
classes from the bootclasspath. That work allowed the presence of
android.test.base in the bootclasspath to be configured at build time
to allow the changes to be tested without affecting the standard
Android builds and avoiding having to repeatedly reapply/revert the
changes that excluded android.test.base from the bootclasspath. That
change has been applied and stuck and no builds change the default by
setting REMOVE_ATB_FROM_BCP=false so we no longer need to support that
capability.

This change removes the build time switch to add
framework-atb-backward-compatibility to the bootclasspath and another
change in the same topic merges those classes into the
framework-minus-apex module. So, while a module has been removed from
the bootclasspath the classes available on it have not changed.

Bug: 184331423
Test: m nothing
Change-Id: I9dadaf8b0c2684bf1983b353bb2acf4f42655e1a